# backend/swagger.yaml
openapi: 3.0.0
info:
  title: comp3900_project
  contact: {}
  version: "1.0"
servers:
  - url: http://localhost:3000
    variables: {}
paths:
...
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    get:
      tags:
        - User
      summary: Get Single User
      operationId: GetSingleUser
      parameters: []
      responses:
        "200":
          description: "Information about specified user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserToken"
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    UserToken:
      type: object
      properties:
        type:
          type: string
          description: The user ID.
          example: admin
        zid:
          type: string
          description: The zid of the user.
          example: z54322
        email:
          type: string
          description: The email of the user.
          example: tjfxr@unsw.edu.au
        name:
          type: string
          description: The name of the user.
          example: tjfxr
        userId:
          type: string
          description: The user ID of the user.
          example: 667be8938e6a2e7ca5124f92
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...